+++ insanic/utils/__init__.py
@@ -1,5 +1,14 @@
+import aiodns
+import asyncio
+
 import socket
 import ujson as json
+
+from aiodns import DNSResolver, error
+from pycares import errno
+
+from insanic.conf import settings
+
 
 
 def force_str(val):
@@ -28,3 +37,21 @@ def get_my_ip():
     finally:
         s.close()
     return ip
+
+
+def get_my_host_by_addr(ip=None):
+    if ip is None:
+        ip = get_my_ip()
+    fut = asyncio.Future()
+
+    def _callback_gethostbyadd(result, errorno):
+        if fut.cancelled():
+            return
+        if errorno is not None:
+            fut.set_exception(error.DNSError(errorno, errno.strerror(errorno)))
+        else:
+            fut.set_result(result)
+
+    _resolver = aiodns.DNSResolver()
+    _resolver._channel.gethostbyaddr(ip, _callback_gethostbyadd)
+    return fut